added a component to hold the quantity drop down menu
